refactor: change validate price from int to num

+++ app/Http/Requests/Maintenance/CreateMaintenanceRequest.php
@@ -32,7 +32,7 @@ class CreateMaintenanceRequest extends FormRequest
     {//device_id , title , price , date ,note
         return [
             'deviceId' => ['required','exists:devices,id'],
-            'price'=> ['required','integer','min:1'],
+            'price'=> ['required','numeric','min:1'],
             'title' => ['required','string'],
             'place' => ['nullable','string'],
             'date'=>['required','string','date_format:Y-m-d'],
